Create decryptionLoopCore_V1.vhd
This is a pipelined implementation of the decryption core. It is one if the first 9 rounds of decryption. The final round should use a core which is almost identical but does not carry out an invMixColun step in the end.
This implementation requires 7 clock cycles between the first encrypted word is input and the first decrypted word is output.
The words should be input in BIG-ENDIAN mode (that is: MOST-SIGNIFICANT-WORD FIRST).
If there is no more data to decrypt, then six padding words should be input at the end so that the last piece of data has the chance to move along the pipeline. The content of these padding words can be whatever as it has no effect on the validity of the expected output data.