Added comments to the decription.
The first 4 words that the module receives should be the key. Key expansion takes about 139 clock cycles, after which the READY_TO_DECRYPT signal is asserted.
Words should be then input one per clock cycle. It takes 69 cycles for a word to exit the pipeline.
If a clock cycle is missed without inputting a word, block re-alignment is required.

For this test bench, the first valid word output occurs at the rising edge at 427ns.